Create settings file base class
