+++ src/ansible_navigator/utils/version_migration/v1_v2_settings_file.py
@@ -3,25 +3,16 @@
 
 from typing import Dict
 
-from ...content_defs import ContentView
-from ...content_defs import SerializationFormat
-from ..ansi import COLOR
-from ..ansi import changed
-from ..ansi import info
 from ..dot_paths import MergeBehaviors
 from ..dot_paths import check_path
 from ..dot_paths import get_with_path
 from ..dot_paths import move_to_path
 from ..dot_paths import place_at_path
-from ..serialize import Loader
-from ..serialize import serialize_write_file
-from ..serialize import yaml
-from .definitions import Migration
 from .definitions import MigrationStep
-from .definitions import MigrationType
+from .settings_file import SettingsFile
 
 
-class V1V2SettingsFile(Migration):
+class V1V2SettingsFile(SettingsFile):
     """The v1 to v2 settings file migration.
 
     The ordering of the functions in this class is important.
@@ -29,65 +20,12 @@ class V1V2SettingsFile(Migration):
     """
 
     name = "Version 1 to Version 2 settings file format migration"
-    migration_type: MigrationType = MigrationType.SETTINGS_FILE
 
     def __init__(self):
         """Initialize the v1 to v2 settings file migration."""
         super().__init__()
         self.content: Dict = {}
-
-    def run(self, *args, **kwargs) -> None:
-        """Perform the v1 to v2 migration.
-
-        :param args: Positional arguments
-        :param kwargs: Keyword arguments
-        """
-        if not self.content:
-            with self.settings_file_path.open("r", encoding="utf-8") as f:
-                try:
-                    self.content = yaml.load(f, Loader=Loader)
-                except (yaml.scanner.ScannerError, yaml.parser.ParserError):
-                    return
-
-        if not self.content:
-            return
-
-        # Check if any of the migrations are needed
-        if self.check:
-            self.run_steps()
-            self.was_needed = self.needed_now
-            return
-
-        # Not check and wasn't needed
-        if not self.was_needed:
-            return
-
-        self.run_steps()
-
-        # Something may have gone wrong
-        if self.needed_now:
-            return
-
-        # Back up the current
-        backup = self.settings_file_path.rename(self.settings_file_path.with_suffix(".v1"))
-        info(color=COLOR, message=f"Backup: {backup}")
-
-        # Write the new file
-        if self.settings_file_path.suffix in (".yml", ".yaml"):
-            serialization_format = SerializationFormat.YAML
-        elif self.settings_file_path.suffix == ".json":
-            serialization_format = SerializationFormat.JSON
-
-        serialize_write_file(
-            content=self.content,
-            content_view=ContentView.NORMAL,
-            file_mode="w",
-            file=self.settings_file_path,
-            serialization_format=serialization_format,
-        )
-        changed(color=COLOR, message=f"Updated: {self.settings_file_path}")
-
-        return
+        self._backup_suffix = ".v1"
 
     @MigrationStep.register(MigrationStep(name="config path"))
     def config_path(self) -> bool:
